Changed location of the __init__.py file and made each test function a separate file
