--- desktop.ini
@@ -1,5 +1,0 @@
-[.ShellClassInfo]
-InfoTip=Esta carpeta se ha compartido online.
-IconFile=C:\Program Files\Google\Drive\googledrivesync.exe
-IconIndex=16
-
