+++ tests/test_dataframe_ndarray.py
@@ -1,0 +1,16 @@
+import pytest
+import numpy as np
+from ie-pandas import dataframe
+
+test_data = {'species':['falcon', 'dog', 'spider', 'fish', 'cat'],
+                'num_legs': [2, 4, 8, 0, 4],
+                'num_wings': [2, 0, 0, 0, 0],
+                'num_specimen_seen': [10, 2, 1, 8, 9],
+                'statistics':[-1.509059, 0.119209, -0.494929, 1.047249, 0.276232],
+                'mammal': [False, True, False, False, True]}
+
+# Check the function df["col_name"] returns NumPy arrays
+def check_df_nparray():
+    test_df = dataframe(test_data)
+    for array in test_df.columns:
+        assert isinstance(test_df[array], np.ndarray)
